postinstall.js creation
